requirements.txt
redis
uvicorn
fastapi
openpyxl
requests
tqdm
psycopg2-binary